requirements.txt
openai==1.30.1
gradio==4.44.0
pypdf==4.2.0
python-dotenv==1.0.1